Added checking user settings and publishing on our side
